+++ build_web_compilers/test/dev_compiler_builder_test.dart
@@ -166,6 +166,49 @@ void main() {
             outputs: expectedOutputs);
       });
 
+      test('emits debug symbols when enabled', () async {
+        var expectedOutputs = {
+          'b|lib/b${jsModuleExtension(soundNullSafety)}': isNotEmpty,
+          'b|lib/b${jsSourceMapExtension(soundNullSafety)}': isNotEmpty,
+          'b|lib/b${metadataExtension(soundNullSafety)}': isNotEmpty,
+          'b|lib/b${symbolsExtension(soundNullSafety)}': isNotEmpty,
+          'a|lib/a${jsModuleExtension(soundNullSafety)}': isNotEmpty,
+          'a|lib/a${jsSourceMapExtension(soundNullSafety)}': isNotEmpty,
+          'a|lib/a${metadataExtension(soundNullSafety)}': isNotEmpty,
+          'a|lib/a${symbolsExtension(soundNullSafety)}': isNotEmpty,
+          'a|web/index${jsModuleExtension(soundNullSafety)}': isNotEmpty,
+          'a|web/index${jsSourceMapExtension(soundNullSafety)}': isNotEmpty,
+          'a|web/index${metadataExtension(soundNullSafety)}': isNotEmpty,
+          'a|web/index${symbolsExtension(soundNullSafety)}': isNotEmpty,
+        };
+        await testBuilder(
+            DevCompilerBuilder(
+                platform: ddcPlatform,
+                emitDebugSymbols: true,
+                soundNullSafety: soundNullSafety),
+            assets,
+            outputs: expectedOutputs);
+      });
+
+      test('does not emit debug symbols by default', () async {
+        var expectedOutputs = {
+          'b|lib/b${jsModuleExtension(soundNullSafety)}': isNotEmpty,
+          'b|lib/b${jsSourceMapExtension(soundNullSafety)}': isNotEmpty,
+          'b|lib/b${metadataExtension(soundNullSafety)}': isNotEmpty,
+          'a|lib/a${jsModuleExtension(soundNullSafety)}': isNotEmpty,
+          'a|lib/a${jsSourceMapExtension(soundNullSafety)}': isNotEmpty,
+          'a|lib/a${metadataExtension(soundNullSafety)}': isNotEmpty,
+          'a|web/index${jsModuleExtension(soundNullSafety)}': isNotEmpty,
+          'a|web/index${jsSourceMapExtension(soundNullSafety)}': isNotEmpty,
+          'a|web/index${metadataExtension(soundNullSafety)}': isNotEmpty,
+        };
+        await testBuilder(
+            DevCompilerBuilder(
+                platform: ddcPlatform, soundNullSafety: soundNullSafety),
+            assets,
+            outputs: expectedOutputs);
+      });
+
       test('strips scratch paths from metadata', () async {
         var expectedOutputs = {
           'b|lib/b${jsModuleExtension(soundNullSafety)}': isNotEmpty,